git-svn-id: svn://svn.r-forge.r-project.org/svnroot/eat@174 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/R/getConquestVersion.R
@@ -3,32 +3,39 @@ getConquestVersion <- function ( path.conquest , asDate = TRUE ) {
 
 		wd <- getwd ()
 		f <- file.path ( wd , "delete.cqc" )
-		write ( "quit;" , f )
+		if ( file.create ( f, showWarnings = TRUE ) ) {
 		
-		f <- normalizePath ( f )
-		path.conquest <- normalizePath ( path.conquest )
-		cmd <- paste ( '"' , path.conquest , '"' , " " , '"' , f , '"' , sep = "" )
-		r <- NULL
-		ow <- getOption ( "warn" )
-		options ( warn = -1 )
-		try ( r <- system ( command = cmd , intern = TRUE ) , silent = TRUE )
-		options ( warn = ow )
-		
-		file.remove ( f )
-		
-		if ( !is.null ( r ) ) {
-		
-				r <- r[1]
-
-				r <- sub ( "ConQuest build: " , "" , r )
-
-				r <- gsub ( "\\s+" , "-" , r )
+				write ( "quit;" , f )
 				
-				if ( asDate ) r <- as.date(r)
+				f <- normalizePath ( f )
+				path.conquest <- normalizePath ( path.conquest )
+				cmd <- paste ( '"' , path.conquest , '"' , " " , '"' , f , '"' , sep = "" )
+				r <- NULL
+				ow <- getOption ( "warn" )
+				options ( warn = -1 )
+				try ( r <- system ( command = cmd , intern = TRUE ) , silent = TRUE )
+				options ( warn = ow )
 				
+				file.remove ( f )
+				
+				if ( !is.null ( r ) ) {
+				
+						r <- r[1]
+
+						r <- sub ( "ConQuest build: " , "" , r )
+
+						r <- gsub ( "\\s+" , "-" , r )
+						
+						if ( asDate ) r <- as.date(r)
+						
+				}
+
+		} else {
+				r <- NULL
 		}
 		
 		return (r)
+
 }
